pyproject.toml
[tool.poetry]
name = "appsec-analytics"
version = "0.2.0"
description = "AppSec metrics ingestion & analytics (SAST/DAST/SCA): Snyk, ZAP, SonarQube"
authors = ["Evgenii Akimov <you@example.com>"]
packages = [{ include = "appsec_analytics", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
httpx = "^0.27.0"
pydantic = "^2.7.0"
pandas = "^2.2.0"
openpyxl = "^3.1.0"
SQLAlchemy = "^2.0.0"
psycopg = {version = "^3.1.0", extras=["binary"]}
fastapi = "^0.115.0"
uvicorn = {version = "^0.30.0", extras=["standard"]}
typer = "^0.12.0"
python-dotenv = "^1.0.1"
tenacity = "^9.0.0"
orjson = "^3.10.0"
APScheduler = "^3.10.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
ruff = "^0.6.0"
black = "^24.8.0"
mypy = "^1.11.0"
pre-commit = "^3.7.0"

[tool.poetry.scripts]
appsec = "appsec_analytics.cli:app"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B","A","C4","T20","SIM","S"]
ignore = ["S101"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
